Par 22 repositories zu entit ten erstellen

+++ backend/src/main/java/com/gfi/parkplatzapp/backend/persistence/entities/Mitarbeiter.java
@@ -17,7 +17,7 @@ public class Mitarbeiter {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int mitarbeiterID;
+    private Long mitarbeiterID;
 
     private String name;
     private String mail;
